Remove default and default_vmm toolchains
`default` is just an alias for gcc_host, so remove `default` and replace
with `gcc_host` in the default configuration. Users can change to
`clang_host` if they like.

`default_vmm` attempted to switch on the target architecture, but could
not because of the way cmake toolchains are loaded. Replace by
explicitly specifying the target-specific toolchain, setting a sane
default in `config/default.cmake`, and adding a build rule to block
explicit use of the wrong toolchain.

Signed-off-by: Chris Pavlina <pavlinac@ainfosec.com>